\name{APIpkg-package}
\alias{APIpkg-package}
\alias{APIpkg}
\docType{package}
\title{
\packageTitle{APIpkg}
}
\description{
\packageDescription{APIpkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{APIpkg}
\packageIndices{APIpkg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{APIpkg}

Maintainer: \packageMaintainer{APIpkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
